requirements.txt
# Core RAG Libraries
langchain>=0.1.8
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
openai>=1.3.9
anthropic>=0.7.11

# Local LLM Support
llama-cpp-python>=0.2.0
huggingface_hub>=0.20.0

# Vector Databases
chromadb>=0.4.20
pinecone-client>=3.0.0
sentence-transformers>=2.2.2

# Retrieval & Search
rank-bm25>=0.2.2
scikit-learn>=1.3.2

# Reranking
torch>=2.1.1
transformers>=4.35.2

# Data Processing
pandas>=2.1.3
numpy>=1.26.2
pypdf>=4.0.1
pdfplumber>=0.10.3
python-docx>=0.8.11

# UI & Visualization
streamlit>=1.28.1
plotly>=5.18.0
altair>=5.1.2

# Evaluation
ragas>=0.1.7

# Development & Utilities
python-dotenv>=1.0.0
jupyter>=1.0.0
ipython>=8.18.1
black>=23.12.0
flake8>=6.1.0
pytest>=7.4.3

# Logging & Monitoring
python-json-logger>=2.0.7
